requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
pillow>=10.0.0
matplotlib>=3.7.1
numpy>=1.24.0
watchdog>=3.0.0
